fix(core): bound the courteous claim loops with a long counter
`CallContext.maxTryClaim()` is a `long`, but both courteous claim loops counted
with an `int`: `CourteousCall.call`'s `for (int i = 0; i < maxTryClaim(); ++i)`
and `CourteousBalancedCall.call`'s `for (int i = 0; ; )` with
`if (++i >= maxTry) break`. Binary numeric promotion widens the counter for the
comparison, so it tops out at `Integer.MAX_VALUE`, wraps to `Integer.MIN_VALUE`,
and the condition is true again.

Any `maxTryClaim` above `int` range therefore never bounds the loop. A caller
asking for a finite three billion tries gets an unbounded wait — the opposite of
the request — and the `forceCall` fallback and the `return null` decline below
the loop are unreachable. Every other loop counter in the hierarchy is already a
`long` (`ComposedCall.get`, `UncheckedBalancedCall.call`), which is what marks
this as an oversight rather than a design.

The wide-value case has no regression test on purpose: reproducing the wrap
needs 2^31 iterations, each sleeping at least a millisecond. `HARDENING.md`
records that, and the small-value bound stays covered by the exact-sleep-count
assertions in CourteousCallTests.

Widening the counter turned PIT's `IncrementsMutator` into a `MathMutator`
(`IINC` becomes `LADD`), and the new mutant reverses the counter into an
unbounded loop. `CourteousCallTests`' TestClock had been missed when the
64-sleep budgets went into ComposedCallTests and BalancedCallTests; giving it
the same budget kills the newcomer outright and takes the audited
`CourteousCall.call` ConditionalsBoundaryMutator with it — a member recorded as
watchdog-only liveness for want of a bounded clock, not for want of an exit.

The default `maxTryClaim` of Long.MAX_VALUE is left alone: an indefinite wait
rather than an overdraw is a defensible reading of "courteous", and callers
wanting a bound set one explicitly.

+++ ravina-core/src/main/java/software/sava/services/core/remote/call/CourteousBalancedCall.java
@@ -24,7 +24,7 @@ final class CourteousBalancedCall<I, R> extends GreedyBalancedCall<I, R> {
     this.next = loadBalancer.withContext();
     final long maxTry = callContext.maxTryClaim();
     TRY_NEXT:
-    for (int i = 0; ; ) {
+    for (long i = 0; ; ) {
       if (this.next.capacityState().tryClaimRequest(callContext)) {
         return call.apply(this.next.item());
       } else {
